requirements.txt
gTTS
praw
Pillow
